requirements.txt
fastmcp>=2.12.0
sqlmodel>=0.0.14
sqlalchemy>=2.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
